requirements.txt
numpy
Pillow
python-telegram-bot